#257 remove references to __int64

+++ src/obrc/BRCLoader.h
@@ -82,7 +82,7 @@ class SymData
     int globalCount;
     int argCount;
     int localCount;
-    __int64 index;
+    long long index;
     unsigned fileOffs;
     void insert(BrowseData* xx) { data.push_back(xx); }
 };
